add auth support for Tron client
